+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+xxxxxxxxxxxxxxxxxxx
+===================
+- Parameter `memory` added to `Apify.call()`.
+
 0.5.37 / 2018-06-07
 ===================
 - `PseudoUrl` class can now contain a template for `Request` object creation and `PseudoUrl.createRequest()` method.