+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-XXX
+0.5.38 / 2018-06-22
 ===================
 - Revereted back - `Apify.getApifyProxyUrl()` accepts again `session` and `groups` options instead of
   `apifyProxySession` and `apifyProxyGroups`